+++ src/backend/job/biz-job/src/main/kotlin/com/tencent/bkrepo/job/batch/base/CompositeMongoDbBatchJob.kt
@@ -3,6 +3,9 @@ package com.tencent.bkrepo.job.batch.base
 import com.tencent.bkrepo.job.config.properties.CompositeJobProperties
 import org.slf4j.LoggerFactory
 
+/**
+ * 由多个[ChildMongoDbBatchJob]组成的任务，主要目的是将类似遍历所有Node表这种耗时任务合并成一个，只需要遍历一次就可以完成所有数据的处理
+ */
 abstract class CompositeMongoDbBatchJob<T>(
     private val properties: CompositeJobProperties
 ) : MongoDbBatchJob<T, CompositeJobContext<T>>(properties) {
