+++ src/backend/job/biz-job/src/main/kotlin/com/tencent/bkrepo/job/batch/node/NodeStatCompositeMongoDbBatchJob.kt
@@ -29,7 +29,6 @@ class NodeStatCompositeMongoDbBatchJob(
 
     override fun createChildJobs(): List<ChildMongoDbBatchJob<Node>> {
         return listOf(
-            FolderOfRepoStatChildJob(properties, mongoTemplate),
             ProjectRepoStatChildJob(properties, mongoTemplate)
         )
     }
