+++ src/backend/opdata/biz-opdata/src/main/kotlin/com/tencent/bkrepo/opdata/config/OpConfiguration.kt
@@ -33,9 +33,6 @@ import org.springframework.context.annotation.Configuration
 @Configuration
 @EnableConfigurationProperties(
     OpProperties::class,
-    OpProjectRepoStatJobProperties::class,
-    OpFolderStatJobProperties::class,
-    OpFileSystemStatJobProperties::class,
     OpEmptyFolderStatJobProperties::class
 )
 class OpConfiguration
