--- src/backend/opdata/biz-opdata/src/main/kotlin/com/tencent/bkrepo/opdata/filesystem/StoragePathStatVisitor.kt
@@ -1,66 +1,0 @@
-/*
- * Tencent is pleased to support the open source community by making BK-CI 蓝鲸持续集成平台 available.
- *
- * Copyright (C) 2022 THL A29 Limited, a Tencent company.  All rights reserved.
- *
- * BK-CI 蓝鲸持续集成平台 is licensed under the MIT license.
- *
- * A copy of the MIT License is included in this file.
- *
- *
- * Terms of the MIT License:
- * ---------------------------------------------------
- * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated
- * documentation files (the "Software"), to deal in the Software without restriction, including without limitation the
- * rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to
- * permit persons to whom the Software is furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all copies or substantial portions of
- * the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT
- * LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN
- * NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY,
- * WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
- * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
- */
-
-package com.tencent.bkrepo.opdata.filesystem
-
-import com.tencent.bkrepo.common.artifact.path.PathUtils
-import com.tencent.bkrepo.opdata.pojo.storage.PathStatMetric
-import java.io.IOException
-import java.nio.file.FileVisitResult
-import java.nio.file.Path
-import java.nio.file.SimpleFileVisitor
-import java.nio.file.attribute.BasicFileAttributes
-
-/**
- * 遍历目录统计其子目录以及文件相关信息
- */
-class StoragePathStatVisitor(
-    private val rootPath: String,
-    private val pathStatMetric: PathStatMetric
-) : SimpleFileVisitor<Path>() {
-    @Throws(IOException::class)
-    override fun visitFile(filePath: Path, attributes: BasicFileAttributes): FileVisitResult {
-        val file = filePath.toFile()
-        pathStatMetric.totalFileCount += 1
-        pathStatMetric.totalSize += file.length()
-        PathUtils.resolveAncestor(filePath.toString()).forEach {
-            if (it.startsWith(rootPath)) {
-                val temp = pathStatMetric.folders[it] ?: 0
-                pathStatMetric.folders[it] = temp.plus(file.length())
-            }
-        }
-        return FileVisitResult.CONTINUE
-    }
-
-    @Throws(IOException::class)
-    override fun postVisitDirectory(dirPath: Path, exc: IOException?): FileVisitResult {
-        if (dirPath.startsWith(rootPath)) {
-            pathStatMetric.totalFolderCount += 1
-        }
-        return FileVisitResult.CONTINUE
-    }
-}
